pyproject.toml
# pyproject.toml
# Minimal, production-lean setup for a tabular ML + DuckDB + FastAPI project.
# Works with: `make setup` (installs editable + dev deps) and `make serve`.

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-risk-system"
version = "0.1.0"
description = "Revenue-driven fraud risk scoring system (DuckDB-first) with reproducible pipelines and serving."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # Core data + ML
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "scikit-learn>=1.4",
  "lightgbm>=4.3",
  "xgboost>=2.0",
  "shap",
  "duckdb>=0.10",

  # Config + CLI
  "pyyaml>=6.0",
  "pydantic>=2.6",

  # Visualization/reporting (optional but helpful early)
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5.0",
  "ipykernel>=6.29",     # optional: notebooks
  "rich>=13.7",          # nicer CLI logs (optional)
  "joblib>=1.3",         # model artifact IO (optional)
  # Serving stack (kept in dev until you start deploying)
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
